requirements.txt
discord.py
Flask
dnspython
pytz
pymongo
Pillow
